+++ example-functional/src/main/java/com/dan323/functional/data/tree/Leaf.java
@@ -1,32 +1,24 @@
 package com.dan323.functional.data.tree;
 
-import java.util.Objects;
-
 public final class Leaf<A> implements BinaryTree<A> {
 
-    private final A a;
-
-    Leaf(A a) {
-        if (a == null) throw new IllegalArgumentException("The data must not be null.");
-        this.a = a;
+    private Leaf() {
     }
 
-    public A value() {
-        return a;
-    }
+    static final Leaf<?> LEAF = new Leaf<>();
 
     @Override
     public int hashCode() {
-        return 85 * a.hashCode();
+        return 85;
     }
 
     @Override
     public boolean equals(Object obj) {
-        if (obj == this) return true;
-        if (obj instanceof Leaf<?> leaf) {
-            return Objects.equals(leaf.a, a);
-        } else {
-            return false;
-        }
+        return obj instanceof Leaf;
+    }
+
+    @Override
+    public String toString() {
+        return "l[]";
     }
 }