+++ functional-definitions/annotation-definitions/src/main/java/com/dan323/functional/annotation/algs/IMonoid.java
@@ -1,6 +1,5 @@
 package com.dan323.functional.annotation.algs;
 
 public interface IMonoid<A> extends ISemigroup<A> {
-
     String UNIT_NAME = "unit";
 }