+++ functional-definitions/annotation-definitions/src/main/java/com/dan323/functional/annotation/algs/IRing.java
@@ -1,0 +1,8 @@
+package com.dan323.functional.annotation.algs;
+
+public interface IRing<A> extends Algebraic {
+
+    String SUM_NAME = "sum";
+    String PROD_NAME = "prod";
+
+}